Voici l'application avec les fonctionnalités de connexion, inscription, filtrage des voyages, détail d'un voyage.
J'ai aussi fait le spinner avec les dates et le nombre de place disponible.

+++ app/src/main/java/com/Alexis_Jimmy_Yasmine/agencetouristique7/modeles/entitees/Client.java
@@ -1,8 +1,8 @@
 package com.alexis_jimmy_yasmine.agencetouristique7.modeles.entitees;
-//
 
 public class Client {
-    private int id;
+
+    private long id;
     private String nom;
     private String prenom;
     private String email;
@@ -14,7 +14,7 @@ public class Client {
     public Client() {
      }
 
-     public Client(int id, String nom, String prenom, String email, String mdp, int age, String telephone, String adresse) {
+     public Client(long id, String nom, String prenom, String email, String mdp, int age, String telephone, String adresse) {
         this.id = id;
         this.nom = nom;
         this.prenom = prenom;
@@ -25,11 +25,11 @@ public class Client {
         this.adresse = adresse;
     }
 
-     public int getId() {
+     public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
